MetadataAPI/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
boto3==1.26.129
python-decouple==3.8
python-jose[cryptography]==3.3.0
pydantic==1.10.7